Parse list of lists in args

+++ twoSum/main.cpp
@@ -9,7 +9,7 @@ std::vector<int> twoSum(std::vector<int> &nums, int target)
 
     std::unordered_map<int, int> hashMap;
 
-    for (int i = 0; i < nums.size(); i++)
+    for (int i = 0; static_cast<uint>(i) < nums.size(); i++)
     {
         int diff = target - nums[i];
